Move base encoding to `convert`
