+++ identity_core/src/convert/base_encoding.rs
