+++ identity_core/src/convert/mod.rs
@@ -1,10 +1,12 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-//! Traits for JSON conversions between types.
+//! Functionality for JSON conversion and Base de- and encoding.
 
 pub use self::json::FmtJson;
 pub use self::json::FromJson;
 pub use self::json::ToJson;
+pub use base_encoding::*;
 
+mod base_encoding;
 mod json;
