--- identity_core/src/utils/mod.rs
@@ -1,8 +1,0 @@
-// Copyright 2020-2022 IOTA Stiftung
-// SPDX-License-Identifier: Apache-2.0
-
-//! Misc. utility functions (encoding, decoding).
-
-mod base_encoding;
-
-pub use self::base_encoding::*;
